correct checkout in script
The script file used a checkout within the same branch, which
caused all changed files to be kept in the index. This caused a subsequent
problem when compiling the fits distribution.

This commit changes the checkout in a manner that the desired version is
put into a detached head and thus replicates the exact state of the desired commit.